Update build system for M-builds

+++ releng/org.eclipse.rse.build/setup.sh
@@ -89,12 +89,18 @@ if [ ! -f eclipse/plugins/org.eclipse.cdt.core_4.0.1.200709241202.jar ]; then
   cd ${CDTTMP}
   unzip ../cdt-master-4.0.1.zip
   cd ..
-  java -jar eclipse/plugins/org.eclipse.equinox.launcher_1.0.1.R33x_v20070828.jar \
-    -application org.eclipse.update.core.standaloneUpdate \
-    -command install \ 
-    -from file://${CDTTMP} \
-    -featureId org.eclipse.cdt.sdk \
-    -version org.eclipse.cdt.sdk_4.0.1.200709241202
+  CMD="java -jar eclipse/plugins/org.eclipse.equinox.launcher_1.0.1.R33x_v20070828.jar"
+  CMD=CMD+" -application org.eclipse.update.core.standaloneUpdate"
+  CMD=CMD+" -command install"
+  CMD=CMD+" -from file://${CDTTMP}"
+  CMD=CMD+" -featureId org.eclipse.cdt.sdk"
+  CMD=CMD+" -version org.eclipse.cdt.sdk_4.0.1.200709241202"
+  ${CMD}
+  rc=$?
+  if [ ${rc} != 0 ]; then
+    echo "${CMD}"
+    exit 1
+  fi  
   rm -rf ${CDTTMP}
   rm cdt-master-4.0.1.zip
 fi
